Adapt to new structure, add roocs-grids main branch, flatten tests, use xesmf v0.7.0, remove notebook warnings
