+++ pyproject.toml
@@ -48,7 +48,7 @@ dynamic = ["description", "version"]
 dependencies = [
   "bottleneck>=1.3.1",
   # cf-xarray is differently named on conda-forge
-  "cf-xarray>=0.7.0",
+  "cf-xarray>=0.7.5",
   "cftime>=1.4.1",
   "dask[complete]>=2.6",
   "geopandas>=0.11",
@@ -94,7 +94,7 @@ docs = [
   "sphinx",
   "sphinx-rtd-theme>=1.0",
 ]
-extra = ["xesmf>=0.6.2"]
+extra = ["xesmf>=0.7.0"]
 
 [project.urls]
 "Homepage" = "https://clisops.readthedocs.io/"
