+++ requirements_upstream.txt
@@ -2,5 +2,6 @@ bottleneck @ git+https://github.com/pydata/bottleneck.git@master
 cf-xarray @ git+https://github.com/xarray-contrib/cf-xarray.git@main
 cftime @ git+https://github.com/Unidata/cftime.git@master
 flox @ git+https://github.com/xarray-contrib/flox.git@main
+roocs-grids @ git+https://github.com/roocs/roocs-grids.git@main
 roocs-utils @ git+https://github.com/roocs/roocs-utils.git@master
 xarray @ git+https://github.com/pydata/xarray.git@main
