+++ tests/test_output_utils.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import tempfile
+import time
 from pathlib import Path
 
 import xarray as xr
@@ -10,6 +11,7 @@ from clisops import CONFIG
 from clisops.utils.common import expand_wildcards
 from clisops.utils.file_namers import get_file_namer
 from clisops.utils.output_utils import (
+    FileLock,
     get_chunk_length,
     get_da,
     get_output,
@@ -271,3 +273,36 @@ def test_unify_chunks_cmip6():
     assert chunked_ds1.chunks == chunked_ds2.chunks
     # test that ds = ds.unify_chunks hasn't changed ds.chunks
     assert chunked_ds2.chunks == chunked_ds2_unified.chunks
+
+
+def test_filelock_simple(tmp_path):
+    LOCK_FILE = Path(tmp_path, "test.lock")
+    DATA_FILE = Path(tmp_path, "test.dat")
+
+    lock = FileLock(LOCK_FILE)
+
+    lock.acquire()
+    try:
+        time.sleep(1)
+        assert os.path.isfile(LOCK_FILE)
+        assert lock.state == "LOCKED"
+        open(DATA_FILE, "a").write("1")
+    finally:
+        lock.release()
+
+    time.sleep(1)
+    assert not os.path.isfile(LOCK_FILE)
+
+
+def test_filelock_already_locked(tmp_path):
+    LOCK_FILE = Path(tmp_path, "test.lock")
+
+    lock1 = FileLock(LOCK_FILE)
+    lock2 = FileLock(LOCK_FILE)
+
+    lock1.acquire(timeout=10)
+
+    try:
+        lock2.acquire(timeout=5)
+    except Exception as exc:
+        assert str(exc) == f"Could not obtain file lock on {LOCK_FILE}"
