+++ tox.ini
@@ -45,6 +45,8 @@ deps =
     coveralls
     pytest-cov
 commands_pre =
+    # FIXME: Needed to install roocs-grids from git; Remove when roocs-grids is on PyPI
+    pip install git+https://github.com/roocs/roocs-grids.git@main#egg=roocs-grids
     pip list
     pip check
 commands =
